made all the routes protected and send the authorization headers with the token from the frontend

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,12 +1,20 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext.jsx";
+import { useAuthContext } from "../hooks/useAuthContext.jsx";
 
 import formatDistanceToNow from "date-fns/formatDistanceToNow"
 
 const WorkoutDetails = ({workout}) => {
   const {dispatchFunctions}=useWorkoutsContext();
+  const {authState}=useAuthContext();
   const handleClick=async () => {
+    if(!authState.user){
+      return
+    }
     const response = await fetch(`http://localhost:3001/api/workouts/${workout._id}`,{
             method:"DELETE",
+            headers:{
+              "Authorization":`Bearer ${authState.user.token}`
+            }
     })
     const json = await response.json();
     if(response.ok){
